requirements.txt
apify-client
python-telegram-bot
python-dotenv
apscheduler
gspread
google-auth
